feat(react-table): convert react-table to typescript
