+++ packages/patternfly-4/react-core/src/components/Dropdown/InternalDropdownItem.d.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-export interface InternalDropdownItemProps extends React.HTMLProps<HTMLAnchorElement> {
+export interface InternalDropdownItemProps extends React.HTMLProps<HTMLDivElement> {
   /** Anything which can be rendered as dropdown item */
   children?: React.ReactNode; 
   /** Classes applied to root element of dropdown item */
@@ -25,5 +25,5 @@ export interface InternalDropdownItemProps extends React.HTMLProps<HTMLAnchorEle
     sendRef?: (index: number, ref: any, isDisabled: boolean) => void
   };
   /** Callback for click event */
-  onClick?: () => void; 
+  onClick?: (event: any) => void;
 }
