+++ packages/patternfly-4/react-table/package.json
@@ -5,6 +5,7 @@
   "main": "dist/js/index.js",
   "module": "dist/esm/index.js",
   "types": "dist/js/index.d.ts",
+  "patternfly:src": "src/",
   "sideEffects": false,
   "publishConfig": {
     "access": "public",
@@ -17,8 +18,7 @@
   "keywords": [
     "react",
     "patternfly",
-    "table",
-    "reacttabular"
+    "table"
   ],
   "author": "Red Hat",
   "license": "MIT",
@@ -42,13 +42,16 @@
     "react-dom": "^15.6.2 || ^16.4.0"
   },
   "scripts": {
-    "build": "yarn build:babel && node ./scripts/copyTS.js && node ./build/copyStyles.js",
+    "build": "yarn build:babel && yarn build:types && node ./scripts/copyTS.js",
     "build:babel": "concurrently 'yarn build:babel:esm && yarn build:babel:umd' 'yarn build:babel:cjs'",
     "build:babel:cjs": "babel --source-maps --extensions '.js,.ts,.tsx' src --out-dir dist/js --presets=@babel/env",
     "build:babel:esm": "babel --source-maps --extensions '.js,.ts,.tsx' src --out-dir dist/esm",
     "build:babel:umd": "babel --source-maps --extensions '.js' dist/esm --out-dir dist/umd --plugins=transform-es2015-modules-umd",
+    "build:types": "tsc -p tsconfig.gen-dts.json",
     "clean": "rimraf dist",
-    "develop": "yarn build:babel:esm --skip-initial-build --watch --verbose"
+    "develop": "yarn build:babel:esm --skip-initial-build --watch --verbose --source-maps",
+    "lint:ts": "tslint -p tsconfig.typecheck.json",
+    "typecheck": "tsc -p tsconfig.typecheck.json"
   },
   "devDependencies": {
     "@babel/cli": "^7.0.0",
@@ -65,6 +68,7 @@
     "enzyme": "3.9.0",
     "fs-extra": "^6.0.1",
     "glob": "^7.1.2",
-    "rimraf": "^2.6.2"
+    "rimraf": "^2.6.2",
+    "typescript": "3.4.5"
   }
 }
