+++ packages/patternfly-4/react-table/src/components/Table/BodyCell.d.ts
@@ -1,0 +1,5 @@
+import { FunctionComponent } from 'react';
+
+declare const BodyCell: FunctionComponent<any>;
+
+export default BodyCell;
