+++ packages/patternfly-4/react-table/src/components/Table/BodyWrapper.d.ts
@@ -1,0 +1,5 @@
+import { Component } from 'react';
+
+declare const BodyWrapper: Component<any>;
+
+export default BodyWrapper;
