--- packages/patternfly-4/react-table/src/components/Table/Header.d.ts
@@ -1,9 +1,0 @@
-import { FunctionComponent, HTMLProps } from 'react';
-
-export interface HeaderProps extends HTMLProps<HTMLTableRowElement> {
-  className?: string;
-}
-
-declare const Header: FunctionComponent<HeaderProps>;
-
-export default Header;
