+++ packages/patternfly-4/react-table/src/components/Table/HeaderCell.d.ts
@@ -1,0 +1,5 @@
+import { FunctionComponent } from 'react';
+
+declare const HeaderCell: FunctionComponent<any>;
+
+export default HeaderCell;
