+++ packages/patternfly-4/react-table/src/components/Table/SortColumn.tsx
@@ -1,0 +1,43 @@
+import * as React from 'react';
+import { LongArrowAltUpIcon, LongArrowAltDownIcon, ArrowsAltVIcon } from '@patternfly/react-icons';
+import { css } from '@patternfly/react-styles';
+import styles from '@patternfly/react-styles/css/components/Table/table';
+
+export enum SortByDirection {
+  asc = 'asc',
+  desc = 'desc'
+};
+
+interface SortColumnProps extends React.HTMLAttributes<HTMLButtonElement> {
+  children?: React.ReactNode;
+  className?: string;
+  isSortedBy?: Boolean;
+  onSort?: Function;
+  sortDirection?: string;
+} 
+
+const SortColumn: React.FunctionComponent<SortColumnProps> = ({
+  children = null,
+  className = '',
+  isSortedBy = false,
+  onSort = null,
+  sortDirection = '',
+  ...props
+} : SortColumnProps ) => {
+  let SortedByIcon;
+  if (isSortedBy) {
+    SortedByIcon = sortDirection === SortByDirection.asc ? LongArrowAltUpIcon : LongArrowAltDownIcon;
+  } else {
+    SortedByIcon = ArrowsAltVIcon;
+  }
+  return (
+    <button {...props} className={css(className)} onClick={event => onSort && onSort(event)}>
+      {children}
+      <span className={css(styles.tableSortIndicator)}>
+        <SortedByIcon />
+      </span>
+    </button>
+  );
+};
+
+export default SortColumn;
