+++ packages/patternfly-4/react-table/src/components/Table/base/index.d.ts
@@ -1,0 +1,9 @@
+export { default as Provider } from './provider';
+export { default as Header } from './header';
+export { default as Body } from './body';
+export { default as BodyRow } from './body-row';
+export { default as evaluateFormatters } from './evaluate-formatters';
+export { default as evaluateTransforms } from './evaluate-transforms';
+export { default as mergeProps } from './merge-props';
+export { default as columnsAreEqual } from './columns-are-equal';
+export { default as resolveRowKey } from './resolve-row-key';
