+++ packages/patternfly-4/react-table/src/components/Table/utils/headerUtils.d.ts
@@ -1,0 +1,2 @@
+declare function calculateColumns(headerRows: any, extra: object): Array<any>;
+export { calculateColumns };
