+++ packages/patternfly-4/react-table/tsconfig.gen-dts.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.base.json",
+  "compilerOptions": {
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "outDir": "dist/js"
+  }
+}
